requirements.txt
openai>=1.75.0
openai-agents>=0.0.11
loguru>=0.7.3
python-dotenv>=1.0.0

# Dev dependencies
black>=24.2.0
flake8>=7.0.0
pytest>=7.3.1